add: Adding a sharedPreference to save connexion status, associated with the CONNEXION_STATUS_KEY key:
As soon as the user logs out, the MainActivity is cosed, the ConnexionActivity is started, and the sharedPreferences called connexion status must be set to false, by calling setSharedPrefConnexionStatus(false). If the user is logged in again, then you should call setSharedPrefConnexionStatus(true).

A function called checkConnexionStatus() was added to check if the user is logged in or not (and if not, the user cannot access the MainActivity, only the ConnexionActivity). This function must be called in the onResume() function, and not in the onCreate().